settime: small changes (datentypwechsel uint->uint8_t) event.h: small changes (datentypwechsel timesetevt uint->uint8_t)

+++ FreeRTOS_AOs/rtos/rtosbase/src/app/events.h
@@ -70,8 +70,8 @@ typedef struct
 typedef struct 
 {
   QEvent super;
-	unsigned int hours;
-	unsigned int min;
+	uint8_t hours;
+	uint8_t min;
 } TimeSetEvt;
